Update Ensemble_Generator.py
The algorithm generates the said quantities for the 3D Ising Model. To perform the same for a general d-dim Ising Model, modify the neighboring_sites function to include all the neighbors in d-dim and change self.state to create a d-dim lattice with all spin values set to +1. In the wolff_cluster_move function, change seed to tuple(rng.integers(0, self.width, d)), d = dimensions. Finally, update the volume to N^d where N is the length of the lattice generated.